+++ src/simulators/iCubSimulation/EyeLidsController.cpp
@@ -41,7 +41,7 @@ void EyeLids::checkPort()
 	{
         Bottle *bot = port.read(false);
 		if (bot!=NULL){
-	        eyeLidsRotation = bot->get(0).asDouble();
+	        eyeLidsRotation = (float)bot->get(0).asDouble();
             bot->clear();
 		    printf("Message received: %s\n",bot->toString().c_str());
         }
